Move cache to top level, to make clear it can be used elsewhere
