+++ cache/writetrough_cache_test.go
